requirements.txt
django==3.0.6
coverage==5.1
django-nose==1.4.6
django-bootstrap4==1.1.1
django-tables2==2.3.1
django-filter==2.2.0